refactor(library): tighten bounded queue, panel registry, persistence flusher
BoundedSampleQueue grows tests covering wrap-around and overflow accounting.
PanelRegistry, PersistenceFlusher, and SessionStore drop redundant guards
and clarify ownership where the percentile/timeline work uncovered overlap
between the aggregator and the storage layer. Obs.cs aligns the public
metric facade with the new GcObserver and TpsFpsObserver entry points.

+++ RimObs.Collector/Storage/SessionStore.cs
@@ -212,7 +212,7 @@ ON CONFLICT(metric_id, canonical) DO UPDATE SET
         foreach (MetricStats metric in metrics) {
             pId.Value = metric.MetricId;
             pName.Value = metric.Name ?? string.Empty;
-            pKind.Value = metric.Kind;
+            pKind.Value = (byte)metric.Kind;
             pUnit.Value = metric.Unit ?? string.Empty;
             upsertMetric.ExecuteNonQuery();
 
@@ -256,7 +256,7 @@ VALUES ($gen, $pause, $hb, $ha, $dur, $ticks, $rate);
 
             foreach (GcEventRecord e in events) {
                 pGen.Value = e.Generation;
-                pPause.Value = e.PauseType;
+                pPause.Value = (byte)e.PauseType;
                 pHb.Value = e.HeapBefore;
                 pHa.Value = e.HeapAfter;
                 pDur.Value = e.DurationMicros;
